+++ .prettierrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+	semi: true,
+	singleQuote: true,
+	printWidth: 120,
+	useTabs: true,
+	endOfLine: 'auto',
+};
